Add files via upload
Fast Fourier Transform (FFT) is a widely used algorithm for computing the discrete Fourier
transform (DFT) of a sequence of complex or real numbers. FFT has many applications in
various fields, such as signal processing, image processing, data compression, cryptography, and
more. However, FFT is also computationally intensive and requires a large amount of memory
and arithmetic operations. Therefore, implementing FFT efficiently and accurately on resourceconstrained devices is a challenging task.
Picorv32 is a small and simple RISC-V CPU core written in Verilog. It supports the RV32IMC
instruction set and can run on FPGA boards. Picorv32 also provides a Processor-Peripheral
Interface (PCPI) that allows external co-processors to extend the functionality of the CPU. PCPI
is a simple and flexible interface that uses a few signals to communicate between the CPU and
the co-processor